move showCase axios request to map component and pass down ShowCaseData as props to showcase component to remove un mounting error
