+++ src/components/Showcase.jsx
@@ -1,8 +1,6 @@
 // SHOWCASE COMPONENT:
 //-------------------------------------------------------------------
-import React, { useEffect, useState } from "react";
-import axios from "axios";
-
+import React from 'react';
 // MUI CSS
 import Box from "@mui/material/Box";
 import CardActions from "@mui/material/CardActions";
@@ -10,24 +8,12 @@ import CardContent from "@mui/material/CardContent";
 import Typography from "@mui/material/Typography";
 import ShowcaseItem from "./ShowcaseItem";
 //-------------------------------------------------------------------
-// API KEY: (references our .env file)
-const api = process.env.REACT_APP_API;
+
 
 // -when called this func gets all the showcase drawings from the DB via express api
 export default function Showcase(props) {
   //console.log(props);
-  const [showcaseData, setshowcaseData] = useState([]);
-  useEffect(() => {
-    const getShowcaseDrawings = async () => {
-      try {
-        const response = await axios.get(`${api}/showcase`);
-        setshowcaseData(response.data);
-      } catch (e) {
-        return console.log(e);
-      }
-    };
-    getShowcaseDrawings();
-  }, []);
+  
 
   return (
     <Box sx={{ minWidth: 275 }}>
@@ -42,7 +28,7 @@ export default function Showcase(props) {
         </CardContent>
         <CardActions>
           {/* map through the get response data and render a new showcaseItem for each value */}
-          {showcaseData.map((drawing) => {
+          {props.showcaseData.map((drawing) => {
             return (
               <React.Fragment key={drawing.id}>
                 <ShowcaseItem
